[flang] Separate definition of class FirConverter into header file
The order of functions has been preserved, the order in the cpp file
matches the order in the header. This was meant to limit the visual
differences between the old and the new files.

Small functions and templates have their definitions in the header
(except `genFIR`).

All `genFIR` functions (except those explicitly listed as no-ops)
are defined in the cpp file.

The semantic changes are
- `FirCoverter` is defined in namespace `Fortran::lower` (was in
  anonymous namespace before).
- Helper classes `IncrementLoopInfo', `ConstructContext`, and
  `TypeInfoConverter` are now nested inside of `FirConverter` (were
  in anonymous namespace before).